social-cli/requirements.txt
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
